feat: Initial commit
This is the first commit of the Technical Assessment repo. It is
based on an open source dbt project created by Christine Berger and
Lauren Benezra from dbt Labs. This project is characterized by
having TERRIBLE architecture and is very much in need of refactoring.

While unpleasant to work in, it does set us up to create a PR that
reflects the types of changes that we would want to make in our
dbt project. :wink: