style: Formated queries to make them more inteligible
I've formatted the queries to make them easier to work with as we setup a reasonable baseline for this project. Ideally, we'll want to preserve some poor habits so we have something to tweak during the PR in addition to the new functionality. I believe this could be a good way to test for scout's rule.