refactor: Move pre-existing lowcost part supplier queries to analyses
It's good practice to have our purpose-built queries in the BI layer so analysts can adjust filters and tailor the queries for specific departments/dashboards. We dont want to lose the queries, however, so I'm placing them in the `analyses` directory.

BREADKING CHANGE: The `EUR_LOWCOST_BRASS_SUPPLIERS` and `UK_Lowcost_Brass_Suplpiers` models are now static, and will be removed at a later time.